requirements.txt
aspose-words==22.5.0

